pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "CLI-programm for find diff two files"
authors = ["AndreyEpp <epp@mayak.help>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest-cov = "^3.0.0"
pyyaml = "^6.0"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pytest = "^7.1.0"
coverage = "^6.4.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
